rename LeaveRoomOwnerDialog to LeaveRoomProhibitedDialog
