+++ src/modules/ui/room-details/LeaveRoomProhibitedDialog.vue
